requirements.txt
# Core Framework
Flask==3.0.3
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-WTF==1.2.1

# Database
mysql-connector-python==9.0.0
SQLAlchemy==2.0.31  # optional, if you plan to use ORM

# Templates / Utilities
Jinja2==3.1.3
Werkzeug==3.0.2
itsdangerous==2.1.2
click==8.1.7

# Deployment
gunicorn==23.0.0